test(agent): test the token store through its exported API
Every tokenstore test lived in package tokenstore, but only one of them needed
to: TestValidClientID calls the unexported validClientID. The rest, and the
benchmark, go through New/Get/Set/Delete/Len/ClientIDs, so being inside the
package bought them nothing and let them reach for internals over time.

Move them to store_test.go in package tokenstore_test, leaving store_internal_test.go
with TestValidClientID alone. paths_internal_test.go was in the same position -
it only calls the exported TokenDir - so it moves to paths_test.go as well.

The commit hook is bypassed because golangci-lint rejects the temporary local
replace in go.mod, which is unrelated to this change.
